requirements.txt
click==7.1.2
future==0.18.2
importlib-metadata==3.1.0
Jinja2==2.11.2
joblib==0.17.0
livereload==2.6.3
lunr==0.5.8
Markdown==3.3.3
MarkupSafe==1.1.1
mkdocs==1.1.2
mkdocs-awesome-pages-plugin==2.4.0
mkdocs-macros-plugin==0.5.0
mkdocs-macros-test==0.1.0
mkdocs-material==6.2.8
mkdocs-material-extensions==1.0.1
mkdocs-redirects==1.0.1
mkdocs-section-index==0.3.0
nltk==3.5
Pillow==8.4.0
Pygments==2.7.2
pymdown-extensions==8.0.1
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.11.13
six==1.15.0
termcolor==1.1.0
tornado==6.1
tqdm==4.54.0
zipp==3.4.0
